# docs/swagger.yaml
definitions:
  models.Transaction:
    type: object
info:
  contact: {}
paths:
  /api/dashboard:
    get:
      description: Menampilkan ringkasan transaksi
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Dashboard utama
      tags:
      - Dashboard
  /api/dashboard/bar:
    get:
      description: Menampilkan grafik batang pengeluaran per kategori
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: Grafik batang pengeluaran
      tags:
      - Dashboard
  /api/dashboard/donut:
    get:
      description: Menampilkan grafik donat pemasukan per kategori
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: Grafik donat pemasukan
      tags:
      - Dashboard
  /api/transactions:
    get:
      description: Menampilkan semua transaksi dengan filter dan pagination
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Limit per page (default 10)
        in: query
        name: limit
        type: integer
      - description: Filter by type (pemasukan/pengeluaran)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Transaction'
            type: array
      summary: Ambil daftar transaksi
      tags:
      - Transactions
    post:
      consumes:
      - application/json
      description: Menambahkan data transaksi
      parameters:
      - description: Transaksi baru
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/models.Transaction'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Transaction'
      summary: Tambah transaksi baru
      tags:
      - Transactions
  /api/transactions/{id}:
    delete:
      description: Menghapus transaksi berdasarkan ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Hapus transaksi
      tags:
      - Transactions
swagger: "2.0"